refactor(#3270): extract funcref-wrapper-types, callback-classification, funcref-as-closure from closures.ts
Behaviour-preserving verbatim extraction (byte-identity IDENTICAL 39/39, tsc 0). Part of #3182 god-file breakdown.

✓
