+++ src/codegen/closures.ts
@@ -18,7 +18,6 @@ import { ts, forEachChild } from "../ts-api.js";
 import { isVoidType, unwrapPromiseType, isPromiseType } from "../checker/type-mapper.js";
 import type { FieldDef, Instr, LocalDef, StructTypeDef, ValType } from "../ir/types.js";
 import { isStandalonePromiseActive } from "./async-scheduler.js"; // (#2867 Gap 1) native-$Promise carrier gate
-import { classMemberFuncKey } from "./class-member-keys.js"; // (#1983) collision-free class-member funcMap keys
 import { definedFuncAt, funcSignatureOf, mintDefinedFunc, pushDefinedFunc } from "./func-space.js"; // (#1916 S2 read chokepoint / S3b stable-regime minting)
 import { inLiveShiftRange } from "../emit/resolve-layout.js"; // (#1916 S3b) manual import-shift must skip stable handles
 import { addStringConstantGlobal } from "./registry/imports.js"; // (#2025)
@@ -94,6 +93,23 @@ import {
   registerNativeGenerator,
 } from "./generators-native.js";
 import type { NativeGeneratorInfo } from "./context/types.js";
+// (#3270) Extracted closure subsystems. Re-exported below so external importers
+// that reference these symbols via `./closures.js` are unaffected.
+import {
+  getFuncSignature,
+  getOrCreateFuncRefWrapperTypes,
+  getFuncRefWrapperRootTypeIdx,
+} from "./closures/funcref-wrapper-types.js";
+export { getFuncSignature, getOrCreateFuncRefWrapperTypes, getFuncRefWrapperRootTypeIdx };
+import {
+  isVecOrArrayRefType,
+  isHostCallbackArgument,
+  isDeferredCallbackArgument,
+  isJsonReviverArgument,
+} from "./closures/callback-classification.js";
+export { isVecOrArrayRefType, isHostCallbackArgument, isDeferredCallbackArgument };
+import { emitFuncRefAsClosure } from "./closures/funcref-as-closure.js";
+export { emitFuncRefAsClosure };
 
 // ── Arrow function callbacks ──────────────────────────────────────────
 
@@ -1261,346 +1277,6 @@ export function emitMethodParamDefaults(
 }
 
 /**
- * #1311 — Host method names whose callable arg ALWAYS needs a JS-callable
- * `__make_callback` externref. These methods invoke the callback during the
- * call itself (the JS-side host implementation calls back into the runtime),
- * so the GC-struct closure shape can't satisfy them.
- *
- * Methods NOT in this set get the closure-struct path when their param is
- * callable — the value is stored, not invoked, so the cast at the eventual
- * dispatch site works. Examples: `Map.set`, `WeakMap.set`, `Set.add`,
- * `Array.push`, `Array.unshift`, user-defined methods.
- *
- * Note: array HOFs (`forEach`, `map`, `filter`, `reduce`, etc.) have
- * dedicated inline compilation in `src/codegen/array-methods.ts` and never
- * reach `isHostCallbackArgument` for their callback arg. They're listed
- * here defensively so that if the inline path is bypassed (e.g. on an
- * untyped receiver), the host-callback path is still chosen.
- */
-const HOST_CALLBACK_METHODS = new Set<string>([
-  // Array HOFs (defensive fallback — usually inlined upstream)
-  "forEach",
-  "map",
-  "filter",
-  "reduce",
-  "reduceRight",
-  "every",
-  "some",
-  "find",
-  "findIndex",
-  "findLast",
-  "findLastIndex",
-  "flatMap",
-  "sort",
-  // Promise prototype methods — JS microtask scheduler invokes the callback
-  "then",
-  "catch",
-  "finally",
-  // Object/JSON callbacks
-  "fromEntries",
-  // String.replace(pattern, replacer) — replacer is a callback
-  "replace",
-  "replaceAll",
-]);
-
-/**
- * (#2070) True when `recvType` is a JS `Array` (`T[]` / `Array<T>`) — used to
- * give `push`/`unshift` callback args the closure-struct path. Recognises the
- * type via its symbol name and, defensively, via the apparent type so a
- * narrowed/aliased array still matches. Typed arrays (`Uint8Array`, …) are not
- * `Array` and correctly fall through to the host-callback default.
- */
-function isArrayLikeReceiverType(recvType: ts.Type, ctx: CodegenContext): boolean {
-  const named = (t: ts.Type | undefined): boolean => t?.getSymbol?.()?.getName?.() === "Array";
-  if (named(recvType)) return true;
-  try {
-    const apparent = ctx.checker.getApparentType?.(recvType);
-    if (named(apparent)) return true;
-    // typeToString covers the structural `T[]` form whose symbol may be absent.
-    const asStr = ctx.checker.typeToString(recvType);
-    if (/(\[\]|^Array<|^ReadonlyArray<)/.test(asStr) || asStr.endsWith("[]")) return true;
-  } catch {
-    // ignore checker errors — default to the host-callback path
-  }
-  return false;
-}
-
-/** (#2903 R4) The nine typed-array view constructor names — a receiver of one
- *  of these types is a native packed-carrier typed array in standalone. */
-const TYPED_ARRAY_VIEW_NAMES: ReadonlySet<string> = new Set([
-  "Int8Array",
-  "Uint8Array",
-  "Uint8ClampedArray",
-  "Int16Array",
-  "Uint16Array",
-  "Int32Array",
-  "Uint32Array",
-  "Float32Array",
-  "Float64Array",
-]);
-
-/** (#2903 R4) Scalar-returning callback HOFs whose standalone typed-array
- *  dispatch is handled natively via `__call_m_*`/`__hof_*` (calls.ts). Excludes
- *  map/filter (typed-RESULT construction — deferred to R4b). */
-const STANDALONE_TYPED_ARRAY_SCALAR_HOFS: ReadonlySet<string> = new Set([
-  "find",
-  "findIndex",
-  "findLast",
-  "findLastIndex",
-  "forEach",
-  "some",
-  "every",
-  "reduce",
-  "reduceRight",
-]);
-
-/**
- * (#2903 R4) True when `recvType` is one of the nine typed-array view types
- * (`Uint8Array`, `Int8Array`, …). Recognises the type via its symbol name and,
- * defensively, via the apparent type so a narrowed/aliased view still matches.
- */
-function isTypedArrayReceiverType(recvType: ts.Type, _ctx: CodegenContext): boolean {
-  // A concrete typed-array receiver carries its view name directly on the type
-  // symbol; the interception is scoped to that (known-element-kind) shape. A
-  // narrowed/aliased view without a direct symbol falls through to the host
-  // path (no regression — that path already worked pre-#2903).
-  const n = recvType.getSymbol?.()?.getName?.();
-  return n !== undefined && TYPED_ARRAY_VIEW_NAMES.has(n);
-}
-
-/**
- * (#2640) True when `wt` is a (nullable) ref to a typed WasmGC vec/array
- * struct (`__vec_*`/`__arr_*`/`$__vec_base`). Used to widen array-like
- * generic-method callback params to externref (see
- * `ctx.forceExternrefCallbackParams`). Mirrors the `__vec_`/`__arr_`
- * receiver-bail detection in `compileArrayLikePrototypeCall`.
- */
-function isVecOrArrayRefType(ctx: CodegenContext, wt: ValType): boolean {
-  if (wt.kind !== "ref" && wt.kind !== "ref_null") return false;
-  const typeIdx = (wt as { typeIdx: number }).typeIdx;
-  const typeDef = ctx.mod.types[typeIdx];
-  const name = typeDef && "name" in typeDef ? (typeDef as { name?: string }).name : undefined;
-  if (!name) return false;
-  return name.startsWith("__vec_") || name.startsWith("__arr_") || name === "__vec_base";
-}
-
-/** Check if an arrow/function expression is used as a callback argument to a call
- *  that targets a HOST import (not a user-defined function). User-defined functions
- *  should receive closures via the GC struct path, not the __make_callback host path. */
-export function isHostCallbackArgument(node: ts.Node, ctx: CodegenContext): boolean {
-  const parent = node.parent;
-  if (!parent) return false;
-  if (ts.isCallExpression(parent)) {
-    if (!parent.arguments.some((arg) => arg === node)) return false;
-    // Check if the callee is a user-defined function — if so, NOT a host callback
-    if (ts.isIdentifier(parent.expression)) {
-      const calleeName = parent.expression.text;
-      const funcIdx = ctx.funcMap.get(calleeName);
-      if (funcIdx !== undefined && funcIdx >= ctx.numImportFuncs) {
-        // User-defined function — use closure path, not host callback
-        return false;
-      }
-      // (#1300) The callee is an identifier but not in funcMap — typically a
-      // function-typed parameter or local. The receiving function expects
-      // the GC-struct closure shape (`__fn_wrap_N_struct`) and will
-      // `ref.cast` the externref it gets. Routing through the host
-      // `__make_callback` path here produces a JS-wrapped externref that
-      // fails the cast and null-derefs at the receiver's `struct.get`.
-      // Detect via TypeScript's call-signature lookup on the identifier's
-      // type and use the closure path if the callee is callable.
-      try {
-        const calleeType = ctx.checker.getTypeAtLocation(parent.expression);
-        const callSigs = calleeType?.getCallSignatures?.();
-        if (callSigs && callSigs.length > 0) {
-          return false;
-        }
-      } catch {
-        // Fall through to host-callback path on any checker error
-      }
-    }
-    // For method calls (property access), check if the method is on a
-    // user-defined class. User-defined methods receive the closure as the
-    // GC-struct shape (`__fn_wrap_N_struct`) and may store it for later
-    // dispatch (e.g. `app.routes.set(path, handler)`). Routing through
-    // `__make_callback` here produces a JS-wrapped externref that fails the
-    // dispatch-site `ref.cast` and null-derefs at `struct.get`. (#1311)
-    if (ts.isPropertyAccessExpression(parent.expression)) {
-      const propAccess = parent.expression;
-      const methodName = propAccess.name.text;
-      // (#3016) `Function.prototype.call`/`apply` NEVER invoke their arguments
-      // as callbacks — they invoke the *receiver* with those args as `thisArg`
-      // + forwarded params. So a function-expression/arrow passed to `.call`/
-      // `.apply` (e.g. `get.call(() => {})` using a function object as an
-      // invalid `this`, or `Array.prototype.find.call(undefined, fn)`) is a
-      // plain function-object VALUE, not a synchronously-invoked host callback.
-      // Routing it through `__make_callback` leaks an `env::` import in
-      // standalone mode for no reason; the GC closure-struct path produces a
-      // valid function-object value host-free (and any HOF that the *receiver*
-      // then invokes — `Array.prototype.forEach.call(arr, cb)` — dispatches the
-      // struct via `__call_fn_N`, verified host-free). Standalone-gated so the
-      // js-host lane stays byte-identical.
-      if (ctx.standalone && (methodName === "call" || methodName === "apply")) {
-        return false;
-      }
-      try {
-        const receiverType = ctx.checker.getTypeAtLocation(propAccess.expression);
-        // Search the receiver type's symbol chain for a class name that
-        // matches a user-defined method `${ClassName}_${methodName}`. We
-        // check both the receiver's own symbol (instance methods) and the
-        // type itself (handles `(typeof Foo).method` for statics).
-        const candidates = new Set<string>();
-        const recSym = receiverType.getSymbol?.();
-        const recName = recSym?.getName?.();
-        if (recName) candidates.add(recName);
-        // (#3231) In standalone/nativeStrings mode a callback passed to a native
-        // DisposableStack `defer`/`adopt` is STORED as a first-class WasmGC closure
-        // (the native runtime later invokes it via `__call_fn_N`), NOT wrapped by
-        // the host `__make_callback`. Route it to the closure-struct path so the
-        // dispatcher can find it host-free. `use` isn't native yet (Phase 1b) — its
-        // callback keeps the host path. Standalone-gated; js-host lane unchanged.
-        if (ctx.nativeStrings && recName === "DisposableStack" && (methodName === "defer" || methodName === "adopt")) {
-          return false;
-        }
-        // Walk base types so inherited user-defined methods are detected
-        const baseTypes = receiverType.getBaseTypes?.();
-        if (baseTypes) {
-          for (const bt of baseTypes) {
-            const bs = bt.getSymbol?.()?.getName?.();
-            if (bs) candidates.add(bs);
-          }
-        }
-        for (const className of candidates) {
-          const fullName = `${className}_${methodName}`;
-          const funcIdx = ctx.funcMap.get(fullName);
-          if (funcIdx !== undefined && funcIdx >= ctx.numImportFuncs) {
-            // User-defined method on a user-defined class — closure path
-            return false;
-          }
-        }
-        // (#2070) Not a user-defined class method. A closure pushed onto an
-        // array via `Array.prototype.push`/`unshift` is *stored*, not invoked,
-        // and the eventual element-read call site (`fns[0]()`) dispatches it as
-        // a WasmGC closure struct. Routing such a closure through the host
-        // `__make_callback` path produces a JS-wrapped externref that fails the
-        // read-site `ref.test`/`ref.cast` and null-derefs at `struct.get`. Give
-        // array storage methods the closure-struct path instead.
-        //
-        // This is deliberately narrow: `Map.set`/`Set.add` and the deferred
-        // DisposableStack methods keep the host-callback path because their
-        // in-class dispatch wrappers (#1311) and writeback machinery (#1695)
-        // depend on the JS-callable externref. The broader
-        // HOST_CALLBACK_METHODS allowlist still governs the invoke-during-call
-        // host methods (array HOFs, Promise.then, String.replace, …).
-        if ((methodName === "push" || methodName === "unshift") && isArrayLikeReceiverType(receiverType, ctx)) {
-          return false;
-        }
-        // (#2903 R4) Standalone typed-array SCALAR-returning callback HOFs
-        // (find/findIndex/…/forEach/some/every/reduce) are dispatched natively
-        // through the `__call_m_<name>`/`__hof_<name>` substrate (see the
-        // interception in `expressions/calls.ts`), which drives the predicate
-        // via `__apply_closure` on a WasmGC closure STRUCT — NOT the host
-        // `__make_callback` externref. Routing the callback through the host
-        // bridge here would leak `env.__make_callback` (breaking host-free
-        // instantiation) AND hand the substrate an externref it can't `ref.cast`
-        // to the closure struct. So give these the closure-struct path.
-        // Standalone-gated (js-host/gc keep the pre-existing host path,
-        // byte-identical). map/filter (typed-RESULT construction) are NOT here —
-        // deferred to R4b. The dyn-view (`$__ta_dyn_view`) receiver shape stays
-        // on its own #3058/#3162 path in array-methods.ts.
-        if (
-          ctx.standalone &&
-          STANDALONE_TYPED_ARRAY_SCALAR_HOFS.has(methodName) &&
-          isTypedArrayReceiverType(receiverType, ctx)
-        ) {
-          return false;
-        }
-      } catch {
-        // Fall through to host-callback path on any checker error
-      }
-    }
-    return true;
-  }
-  // NewExpression: `new Promise(executor)`, `new Map(comparator)`, etc.
-  // Function args to constructors of extern classes need to be JS-callable.
-  if (ts.isNewExpression(parent)) {
-    if (!parent.arguments?.some((arg) => arg === node)) return false;
-    // Check if the constructor is a user-defined class — if so, NOT a host callback
-    if (ts.isIdentifier(parent.expression)) {
-      const ctorName = parent.expression.text;
-      const newFuncIdx = ctx.funcMap.get(classMemberFuncKey(ctx, `${ctorName}_new`)); // (#1983)
-      if (newFuncIdx !== undefined && newFuncIdx >= ctx.numImportFuncs) {
-        return false;
-      }
-      // (#28) `new Promise(executor)` — the executor must be invoked by the host
-      // `Promise_new` import (which does `new Promise(_maybeWrapCallable(executor,
-      // 2, …))`). The `__make_callback` host-callback path does NOT round-trip
-      // here: an INLINE executor (`new Promise((res, rej) => …)`) routed through
-      // it produced no callable wrapper, so the executor was silently never
-      // invoked (resolve/reject `undefined`). Compiling the executor as a
-      // first-class CLOSURE instead emits the `__call_fn_2` dispatcher that
-      // `_maybeWrapCallable` uses to make the wasm closure JS-callable — the same
-      // path the working `const exec = …; new Promise(exec)` form already takes.
-      // So treat the Promise executor as a closure value, not a host callback.
-      if (ctorName === "Promise") {
-        return false;
-      }
-    }
-    return true;
-  }
-  return false;
-}
-
-/**
- * Methods that STORE the callback for later invocation rather than calling it
- * synchronously during the call. Closures passed to these methods need
- * persistent ref-cell writebacks (re-emitted after every subsequent call) so
- * that mutations made when the callback eventually runs are reflected in the
- * outer scope. (#1695)
- *
- * Receiver-type-aware allowlist (className → method names): we only promote
- * to persistent writebacks when the receiver type matches — e.g. a user-defined
- * `class Foo { defer(cb) {} }` calling `foo.defer(...)` must NOT be promoted.
- */
-const DEFERRED_CALLBACK_METHODS_BY_CLASS: ReadonlyMap<string, ReadonlySet<string>> = new Map([
-  ["DisposableStack", new Set(["defer", "use", "adopt"])],
-  ["AsyncDisposableStack", new Set(["defer", "use", "adopt"])],
-]);
-
-/**
- * Returns true if the arrow's parent CallExpression is a stored-callback host
- * method (DisposableStack.defer/use/adopt etc.). The callback is not invoked
- * synchronously by the call that registers it, so its captured-mutable
- * writebacks must be persistent. (#1695)
- */
-export function isDeferredCallbackArgument(node: ts.Node, ctx: CodegenContext): boolean {
-  const parent = node.parent;
-  if (!parent || !ts.isCallExpression(parent)) return false;
-  if (!parent.arguments.some((arg) => arg === node)) return false;
-  if (!ts.isPropertyAccessExpression(parent.expression)) return false;
-  const methodName = parent.expression.name.text;
-  try {
-    const recType = ctx.checker.getTypeAtLocation(parent.expression.expression);
-    const symName = recType.getSymbol?.()?.getName?.();
-    if (symName) {
-      const methods = DEFERRED_CALLBACK_METHODS_BY_CLASS.get(symName);
-      if (methods?.has(methodName)) return true;
-    }
-    const baseTypes = recType.getBaseTypes?.();
-    if (baseTypes) {
-      for (const bt of baseTypes) {
-        const bn = bt.getSymbol?.()?.getName?.();
-        if (!bn) continue;
-        const m = DEFERRED_CALLBACK_METHODS_BY_CLASS.get(bn);
-        if (m?.has(methodName)) return true;
-      }
-    }
-  } catch {
-    // checker failure → conservative false (no behavioural change)
-  }
-  return false;
-}
-
-/**
  * #1177: Returns true if the closure (`arrow`) is provably constructed AFTER
  * the let/const/using declaration of `name` AND the closure is NOT inside a
  * loop that wraps the declaration. In that case, we don't need to force-box
@@ -1876,31 +1552,6 @@ export function genBodyReferencesThis(node: ts.Node): boolean {
     if (!found && genBodyReferencesThis(child)) found = true;
   });
   return found;
-}
-
-/**
- * (#3046) True when `node` is the **reviver** argument (2nd arg) of a
- * `JSON.parse(text, reviver)` call. Per ECMA-262 §25.5.1.1
- * `InternalizeJSONProperty`, the reviver is invoked as
- * `Call(reviver, holder, «name, val»)` — `this` MUST be the holder. The host
- * `JSON_parse` / `_invokeJsonCallable` bridge applies the holder as the JS
- * receiver, so the reviver callback must route through the `this`-forwarding
- * `__make_getter_callback` maker (needsThis) rather than the bare
- * `__make_callback`, which drops the receiver and leaves `this` non-object
- * (a `this.`-op such as `Object.defineProperty(this, …)` then throws
- * "called on non-object").
- */
-function isJsonReviverArgument(node: ts.Node): boolean {
-  const parent = node.parent;
-  if (!parent || !ts.isCallExpression(parent)) return false;
-  if (parent.arguments[1] !== node) return false; // must be the 2nd arg
-  const callee = parent.expression;
-  return (
-    ts.isPropertyAccessExpression(callee) &&
-    ts.isIdentifier(callee.expression) &&
-    callee.expression.text === "JSON" &&
-    callee.name.text === "parse"
-  );
 }
 
 /**
@@ -3850,486 +3501,6 @@ export function compileArrowAsCallback(
 
   fctx.body.push({ op: "call", funcIdx: makeCallbackIdx });
   return { kind: "externref" };
-}
-
-/**
- * Look up a function's parameter and result types from its index.
- */
-export function getFuncSignature(
-  ctx: CodegenContext,
-  funcIdx: number,
-): { params: ValType[]; results: ValType[] } | null {
-  // #1916 S2 — funcSignatureOf is the positional-read chokepoint (func-space.ts).
-  const sig = funcSignatureOf(ctx, funcIdx);
-  return sig ? { params: sig.params, results: sig.results } : null;
-}
-
-/**
- * Get or create the closure struct type and lifted func type for wrapping
- * plain functions with a given signature. Struct type and func type are shared
- * across all functions with the same signature, but each function gets its own
- * trampoline.
- */
-export function getOrCreateFuncRefWrapperTypes(
-  ctx: CodegenContext,
-  userParams: ValType[],
-  resultTypes: ValType[],
-): { structTypeIdx: number; liftedFuncTypeIdx: number; closureInfo: ClosureInfo } | null {
-  // Build cache key from param types and result types
-  const sigKey = `${userParams.map((p) => p.kind + ((p as any).typeIdx ?? "")).join(",")}->${resultTypes.map((r) => r.kind + ((r as any).typeIdx ?? "")).join(",")}`;
-
-  const cached = ctx.funcRefWrapperCache.get(sigKey);
-  if (cached) {
-    return { structTypeIdx: cached.structTypeIdx, liftedFuncTypeIdx: cached.funcTypeIdx, closureInfo: cached };
-  }
-
-  // Create the closure struct type: just (field $func funcref), no captures.
-  // Mark as non-final (superTypeIdx = -1) so closures with captures can be
-  // subtypes of this wrapper struct, enabling ref.cast to succeed at call sites.
-  const closureName = `__fn_wrap_${ctx.closureCounter++}`;
-  const structFields = [{ name: "func", type: { kind: "funcref" as const }, mutable: false }];
-  const structTypeIdx = ctx.mod.types.length;
-  const rootWrapperTypeIdx = (ctx as unknown as { __funcRefWrapperRootTypeIdx?: number }).__funcRefWrapperRootTypeIdx;
-  ctx.mod.types.push({
-    kind: "struct",
-    name: `${closureName}_struct`,
-    fields: structFields,
-    superTypeIdx: rootWrapperTypeIdx ?? -1, // first wrapper is the root; later signatures subtype it
-  });
-  if (rootWrapperTypeIdx === undefined) {
-    (ctx as unknown as { __funcRefWrapperRootTypeIdx?: number }).__funcRefWrapperRootTypeIdx = structTypeIdx;
-  }
-
-  // Create the lifted function type: (ref $struct, ...userParams) -> results
-  const liftedParams: ValType[] = [{ kind: "ref", typeIdx: structTypeIdx }, ...userParams];
-  const liftedFuncTypeIdx = addFuncType(ctx, liftedParams, resultTypes, `${closureName}_type`);
-
-  const closureInfo: ClosureInfo = {
-    structTypeIdx,
-    funcTypeIdx: liftedFuncTypeIdx,
-    returnType: resultTypes.length > 0 ? resultTypes[0]! : null,
-    paramTypes: userParams,
-  };
-  ctx.closureInfoByTypeIdx.set(structTypeIdx, closureInfo);
-  ctx.funcRefWrapperCache.set(sigKey, closureInfo);
-
-  return { structTypeIdx, liftedFuncTypeIdx, closureInfo };
-}
-
-/**
- * (#2873 park fix) The ROOT funcref-wrapper struct type — the FIRST wrapper
- * `getOrCreateFuncRefWrapperTypes` created in this module. Every later
- * per-signature wrapper struct is a `sub final` of it (see the star chaining
- * above), so the root is the ONLY wrapper type a `ref.test`/`ref.cast` is
- * guaranteed to accept for a closure value of ANY signature's wrapper.
- *
- * Why callers need it: wrapper structs are all layout-identical
- * `(struct (field funcref))`, but WasmGC isorecursive canonicalization keys on
- * (fields, supertype, finality) — a `sub final $root` sibling does NOT
- * canonicalize with the root or with another sibling. A call site that casts a
- * closure value to the wrapper of its *declared* signature therefore nulls out
- * whenever the value was allocated under a different signature's wrapper
- * (e.g. an activated async closure: its wrapper is minted for the REWRITTEN
- * `... -> externref` Promise signature, while an `fn: () => void` param casts
- * to the void wrapper) — unless creation ORDER happened to make the declared
- * wrapper the root. Cast to the root instead and discriminate on the funcref's
- * exact type (which encodes the true signature).
- */
-export function getFuncRefWrapperRootTypeIdx(ctx: CodegenContext): number | undefined {
-  return (ctx as unknown as { __funcRefWrapperRootTypeIdx?: number }).__funcRefWrapperRootTypeIdx;
-}
-
-/**
- * (#2976) Emit the memoized, `ref.is_null`-guarded VALUE instance of a
- * capture-carrying nested function declaration:
- *
- *   local.get $memo
- *   ref.is_null
- *   if (empty)                       ;; first DYNAMIC reference only
- *     ref.func $tramp
- *     <capture pushes>               ;; unchanged from the per-site build
- *     struct.new $__fn_cap_<name>
- *     local.set $memo
- *   end
- *   local.get $memo
- *   ref.as_non_null
- *
- * The memo local is allocated once per enclosing activation
- * (`fctx.nestedFnClosureMemos`), so every reference yields the SAME struct
- * instance — `f === f` holds and sidecar/static writes (`f.resolve = fn`)
- * are visible through later references. The runtime guard (not a prologue
- * hoist, not compile-order memoization) is load-bearing twice over:
- *   - it preserves value-capture semantics — immutable captures copy their
- *     value at the first DYNAMIC reference, exactly where the old per-site
- *     build copied them (a prologue hoist would run before hoisted-over
- *     initializers);
- *   - it is control-flow-safe — with compile-order memoization, a reference
- *     in a runtime-skipped branch would leave a later branch reading an
- *     uninitialized local.
- * The capture-push block keeps its compile-time side effects (mutable-capture
- * boxing + localMap rebind, TDZ flag boxing) — they now occur while compiling
- * the guard arm, same net effect as before.
- */
-function emitMemoizedNestedFnClosure(
-  ctx: CodegenContext,
-  fctx: FunctionContext,
-  funcName: string,
-  structTypeIdx: number,
-  trampolineFuncIdx: number,
-  nestedCaptures: NonNullable<ReturnType<CodegenContext["nestedFuncCaptures"]["get"]>>,
-  tdzFlaggedNested: NonNullable<ReturnType<CodegenContext["nestedFuncCaptures"]["get"]>>,
-): void {
-  const numCaptures = nestedCaptures.length;
-  const numTdzFlags = tdzFlaggedNested.length;
-
-  let memoLocal = fctx.nestedFnClosureMemos?.get(funcName);
-  if (memoLocal === undefined) {
-    memoLocal = allocLocal(fctx, `__fnmemo_${funcName}_${fctx.locals.length}`, {
-      kind: "ref_null",
-      typeIdx: structTypeIdx,
-    });
-    (fctx.nestedFnClosureMemos ??= new Map()).set(funcName, memoLocal);
-  }
-
-  fctx.body.push({ op: "local.get", index: memoLocal });
-  fctx.body.push({ op: "ref.is_null" });
-
-  // Build the construction sequence into the guard's then-arm.
-  const savedBody = pushBody(fctx);
-
-  // struct.new fields: func, cap0, cap1, ..., __tdz_*...
-  fctx.body.push({ op: "ref.func", funcIdx: trampolineFuncIdx });
-  // (#1312) Self-reference inside the lifted body of `funcName` itself —
-  // e.g. `function next() { return call(next); }`. The captures are
-  // already in scope as the leading params [0..numCaptures-1] of the
-  // lifted fn (mutable captures arrive as boxed ref cells, immutable as
-  // raw values). We re-push them by param index instead of trying to
-  // dereference `cap.outerLocalIdx`, which points into a different
-  // (outer) scope and yields garbage / null when reused inside the
-  // current lifted body.
-  const isSelfRef = fctx.name === funcName;
-  for (let i = 0; i < nestedCaptures.length; i++) {
-    const cap = nestedCaptures[i]!;
-    if (isSelfRef) {
-      // Captures arrive at param index `i` in the lifted fn (#1312).
-      fctx.body.push({ op: "local.get", index: i });
-      continue;
-    }
-    // (#2029 family A) Cross-fctx capture sourcing. `cap.outerLocalIdx` is a
-    // slot in the function that DECLARED the nested fn; when this
-    // materialization runs inside a DIFFERENT function (an object-literal
-    // accessor body — the enclosing fn's locals are unreachable), baking it
-    // emit-crashes ("local index out of range") or silently reads the wrong
-    // local. `promoteAccessorCapturesToGlobals` promotes such captures to
-    // module globals (shared ref-cell box for mutable, value global for
-    // immutable); prefer those whenever the current fctx cannot resolve the
-    // name itself. Guarded on localMap-absence so owner-fctx behavior is
-    // unchanged (see the #1177 revert note in calls.ts for why a blanket
-    // localMap-first lookup is NOT safe).
-    const capUnresolvedHere = fctx.localMap.get(cap.name) === undefined;
-    if (cap.mutable && cap.valType) {
-      const refCellTypeIdx = getOrRegisterRefCellType(ctx, cap.valType);
-      const boxGlobal = capUnresolvedHere ? ctx.capturedBoxGlobals?.get(cap.name) : undefined;
-      if (fctx.boxedCaptures?.has(cap.name)) {
-        const currentLocalIdx = fctx.localMap.get(cap.name)!;
-        fctx.body.push({ op: "local.get", index: currentLocalIdx });
-      } else if (boxGlobal !== undefined) {
-        // Shared ref-cell box promoted to a module global — live
-        // write-through semantics with the declaring function.
-        fctx.body.push({ op: "global.get", index: boxGlobal.globalIdx });
-        fctx.body.push({ op: "ref.as_non_null" });
-      } else if (capUnresolvedHere && ctx.capturedGlobals.has(cap.name)) {
-        // Value global (the capture is also directly referenced by the
-        // accessor body) — box a copy. Best-effort: writes through the
-        // closure do not propagate back, but the previous behavior was an
-        // out-of-scope local read (emit crash / wrong local).
-        fctx.body.push({ op: "global.get", index: ctx.capturedGlobals.get(cap.name)! });
-        if (ctx.capturedGlobalsWidened.has(cap.name)) {
-          fctx.body.push({ op: "ref.as_non_null" });
-        }
-        fctx.body.push({ op: "struct.new", typeIdx: refCellTypeIdx });
-      } else {
-        // Stage 1 localMap-first lookup reverted — see calls.ts comment.
-        fctx.body.push({ op: "local.get", index: cap.outerLocalIdx });
-        fctx.body.push({ op: "struct.new", typeIdx: refCellTypeIdx });
-        const boxedLocalIdx = allocLocal(fctx, `__boxed_${cap.name}`, {
-          kind: "ref",
-          typeIdx: refCellTypeIdx,
-        });
-        fctx.body.push({ op: "local.tee", index: boxedLocalIdx });
-        fctx.localMap.set(cap.name, boxedLocalIdx);
-        if (!fctx.boxedCaptures) fctx.boxedCaptures = new Map();
-        fctx.boxedCaptures.set(cap.name, { refCellTypeIdx, valType: cap.valType });
-      }
-    } else if (capUnresolvedHere && ctx.capturedGlobals.has(cap.name)) {
-      // (#2029 family A) Immutable capture promoted to a value global by
-      // the accessor-capture pass — read it instead of the out-of-scope
-      // declaring-function local slot.
-      fctx.body.push({ op: "global.get", index: ctx.capturedGlobals.get(cap.name)! });
-      if (ctx.capturedGlobalsWidened.has(cap.name)) {
-        fctx.body.push({ op: "ref.as_non_null" });
-      }
-    } else {
-      fctx.body.push({ op: "local.get", index: cap.outerLocalIdx });
-    }
-  }
-  // #1205 Stage 3: after all value captures, push the boxed TDZ flag refs
-  // (one per TDZ-flagged capture). Sourcing rules mirror calls.ts — see
-  // the FNDECL-A4 cap-prepend block there for the full rationale. The
-  // short version: only trust the LIVE `fctx.tdzFlagLocals[name]` lookup
-  // when it points to an i32 in the current fctx. Otherwise (block-shadow
-  // or cross-fctx transitive) push `i32.const 1` (treat as initialized) —
-  // matches pre-#1205 behavior where the lifted body had no flag check.
-  if (numTdzFlags > 0) {
-    const i32RefCellTypeIdxForFlags = getOrRegisterRefCellType(ctx, { kind: "i32" });
-    for (let ti = 0; ti < tdzFlaggedNested.length; ti++) {
-      const cap = tdzFlaggedNested[ti]!;
-      if (isSelfRef) {
-        // (#1312) Self-reference inside the lifted body — the TDZ-flag
-        // boxed refs arrive as params at index `numCaptures + ti` (after
-        // all value captures). Re-push from there.
-        fctx.body.push({ op: "local.get", index: numCaptures + ti });
-        continue;
-      }
-      const existingBox = fctx.boxedTdzFlags?.get(cap.name);
-      if (existingBox) {
-        fctx.body.push({ op: "local.get", index: existingBox.localIdx });
-      } else {
-        const liveFlagIdx = fctx.tdzFlagLocals?.get(cap.name);
-        const liveType = liveFlagIdx !== undefined ? getLocalType(fctx, liveFlagIdx) : undefined;
-        const liveOk = liveType?.kind === "i32";
-        if (liveOk && liveFlagIdx !== undefined) {
-          fctx.body.push({ op: "local.get", index: liveFlagIdx });
-          fctx.body.push({ op: "struct.new", typeIdx: i32RefCellTypeIdxForFlags });
-        } else {
-          fctx.body.push({ op: "i32.const", value: 1 });
-          fctx.body.push({ op: "struct.new", typeIdx: i32RefCellTypeIdxForFlags });
-        }
-        const flagBoxLocal = allocLocal(fctx, `__tdz_box_${cap.name}`, {
-          kind: "ref",
-          typeIdx: i32RefCellTypeIdxForFlags,
-        });
-        fctx.body.push({ op: "local.tee", index: flagBoxLocal });
-        if (liveOk) {
-          if (!fctx.boxedTdzFlags) fctx.boxedTdzFlags = new Map();
-          fctx.boxedTdzFlags.set(cap.name, {
-            refCellTypeIdx: i32RefCellTypeIdxForFlags,
-            localIdx: flagBoxLocal,
-          });
-          if (!fctx.tdzFlagLocals) fctx.tdzFlagLocals = new Map();
-          fctx.tdzFlagLocals.set(cap.name, flagBoxLocal);
-        }
-      }
-    }
-  }
-  fctx.body.push({ op: "struct.new", typeIdx: structTypeIdx });
-  fctx.body.push({ op: "local.set", index: memoLocal });
-
-  const thenArm = fctx.body;
-  popBody(fctx, savedBody);
-
-  fctx.body.push({ op: "if", blockType: { kind: "empty" }, then: thenArm, else: [] });
-  fctx.body.push({ op: "local.get", index: memoLocal });
-  fctx.body.push({ op: "ref.as_non_null" });
-}
-
-/**
- * Emit a closure struct wrapping a plain function. Creates a per-function
- * trampoline that delegates to the original function.  Struct types are shared
- * across functions with the same signature so they can be reassigned.
- * Pushes the closure struct ref onto the stack and returns its type.
- */
-export function emitFuncRefAsClosure(
-  ctx: CodegenContext,
-  fctx: FunctionContext,
-  funcName: string,
-  funcIdx: number,
-): ValType | null {
-  const sig = getFuncSignature(ctx, funcIdx);
-  if (!sig) return null;
-
-  const nestedCaptures = ctx.nestedFuncCaptures.get(funcName);
-  if (nestedCaptures && nestedCaptures.length > 0) {
-    // Functions with captures: create a closure struct that stores the capture values.
-    // The trampoline extracts captures from the struct and passes them to the original function. (#857)
-    //
-    // (#2976) IDENTITY: the struct type + trampoline are minted ONCE per
-    // funcName (module-level `nestedFnClosureArtifacts` dedupe below), and the
-    // INSTANCE is memoized per enclosing activation in a `ref.is_null`-guarded
-    // local (`fctx.nestedFnClosureMemos`). Previously every reference site
-    // built a fresh struct type + trampoline + instance, so
-    // `Constructor === Constructor` was false and a static/sidecar write
-    // (`Constructor.resolve = fn`) landed on a dead instance the next
-    // reference never saw (the #2671 Promise capability sub-bucket). The
-    // lazy guard — rather than a prologue hoist — preserves the existing
-    // value-capture semantics exactly: immutable captures copy their value at
-    // the FIRST DYNAMIC reference, the same point the old per-site build
-    // copied them; mutable captures were already live through ref cells.
-    const numCaptures = nestedCaptures.length;
-    // #1205 Stage 3: TDZ-flag captures get extra ref-cell fields after the
-    // value captures, mirroring the leading-param layout of the lifted fn.
-    const tdzFlaggedNested = nestedCaptures.filter((c) => c.hasTdzFlag);
-    const numTdzFlags = tdzFlaggedNested.length;
-    // The lifted fn's signature is [valueCaps..., tdzFlagBoxes..., userParams...].
-    const userParams = sig.params.slice(numCaptures + numTdzFlags);
-    const results = sig.results;
-
-    const wrapperTypes = getOrCreateFuncRefWrapperTypes(ctx, userParams, results);
-    if (!wrapperTypes) return null;
-
-    const cachedArtifacts = ctx.nestedFnClosureArtifacts?.get(funcName);
-    if (cachedArtifacts) {
-      const trampIdx = ctx.funcMap.get(cachedArtifacts.trampolineName);
-      if (trampIdx !== undefined) {
-        emitMemoizedNestedFnClosure(
-          ctx,
-          fctx,
-          funcName,
-          cachedArtifacts.structTypeIdx,
-          trampIdx,
-          nestedCaptures,
-          tdzFlaggedNested,
-        );
-        return { kind: "ref", typeIdx: cachedArtifacts.structTypeIdx };
-      }
-    }
-
-    // Create a custom struct with func + capture fields + TDZ-flag fields
-    // (subtype of the base wrapper).
-    const captureFields: FieldDef[] = nestedCaptures.map((_cap, i) => {
-      const capParamType = sig.params[i]!;
-      return { name: `cap${i}`, type: capParamType, mutable: false };
-    });
-    // #1205 Stage 3: append TDZ-flag ref-cell fields after the value captures
-    // so the trampoline's struct.get of the flag uses the correct field index.
-    let i32RefCellTypeIdxForFlags = -1;
-    if (numTdzFlags > 0) {
-      i32RefCellTypeIdxForFlags = getOrRegisterRefCellType(ctx, { kind: "i32" });
-      for (const cap of tdzFlaggedNested) {
-        captureFields.push({
-          name: `__tdz_${cap.name}`,
-          type: { kind: "ref" as const, typeIdx: i32RefCellTypeIdxForFlags },
-          mutable: false,
-        });
-      }
-    }
-    const closureName = `__fn_cap_${funcName}_${ctx.closureCounter++}`;
-    const structTypeIdx = ctx.mod.types.length;
-    ctx.mod.types.push({
-      kind: "struct",
-      name: `${closureName}_struct`,
-      fields: [{ name: "func", type: { kind: "funcref" as const }, mutable: false }, ...captureFields],
-      superTypeIdx: wrapperTypes.structTypeIdx,
-    });
-
-    // Use the base wrapper's func type so call_ref works via subtype cast
-    const liftedFuncTypeIdx = wrapperTypes.liftedFuncTypeIdx;
-
-    const trampolineName = `__fn_tramp_${funcName}_${ctx.closureCounter++}`;
-    const trampolineBody: Instr[] = [];
-    const trampolineLocals: { name: string; type: ValType }[] = [];
-
-    // We always need the casted-self local when we have either >1 value captures
-    // OR any TDZ-flag fields, because each requires a separate `struct.get`.
-    const totalCapFields = numCaptures + numTdzFlags;
-    if (totalCapFields > 1) {
-      trampolineLocals.push({ name: "__casted_self", type: { kind: "ref", typeIdx: structTypeIdx } });
-    }
-    const castedSelfLocal = 1 + userParams.length;
-
-    // Cast self from base struct to custom struct to access capture fields
-    trampolineBody.push({ op: "local.get", index: 0 });
-    trampolineBody.push({ op: "ref.cast", typeIdx: structTypeIdx });
-
-    if (totalCapFields === 1) {
-      // Exactly one capture field (a value capture; TDZ-flag-only with zero
-      // value captures is impossible because each flag is paired with a value).
-      trampolineBody.push({ op: "struct.get", typeIdx: structTypeIdx, fieldIdx: 1 });
-    } else {
-      trampolineBody.push({ op: "local.set", index: castedSelfLocal });
-      // Push value captures first, then TDZ-flag captures, mirroring the
-      // lifted fn's leading-param order.
-      for (let i = 0; i < totalCapFields; i++) {
-        trampolineBody.push({ op: "local.get", index: castedSelfLocal });
-        trampolineBody.push({ op: "struct.get", typeIdx: structTypeIdx, fieldIdx: i + 1 });
-      }
-    }
-    for (let i = 0; i < userParams.length; i++) {
-      trampolineBody.push({ op: "local.get", index: i + 1 });
-    }
-    trampolineBody.push({ op: "call", funcIdx });
-
-    const trampolineFuncIdx = mintDefinedFunc(ctx);
-    pushDefinedFunc(ctx, trampolineFuncIdx, {
-      name: trampolineName,
-      typeIdx: liftedFuncTypeIdx,
-      locals: trampolineLocals,
-      body: trampolineBody,
-      exported: false,
-    });
-    ctx.funcMap.set(trampolineName, trampolineFuncIdx);
-
-    // Register closureInfo so array method callbacks can use call_ref
-    const closureInfo: ClosureInfo = {
-      structTypeIdx,
-      funcTypeIdx: wrapperTypes.closureInfo.funcTypeIdx,
-      returnType: results.length > 0 ? results[0]! : null,
-      paramTypes: userParams,
-    };
-    ctx.closureInfoByTypeIdx.set(structTypeIdx, closureInfo);
-
-    // (#2976) Register the module-level artifacts so every later reference —
-    // in this or any other fctx — reuses this ONE struct type + trampoline
-    // instead of minting fresh ones per site. Stored by trampoline NAME
-    // (re-resolved via funcMap at emission) so late-import shifts can't
-    // desync a cached raw index.
-    (ctx.nestedFnClosureArtifacts ??= new Map()).set(funcName, { structTypeIdx, trampolineName });
-
-    emitMemoizedNestedFnClosure(
-      ctx,
-      fctx,
-      funcName,
-      structTypeIdx,
-      trampolineFuncIdx,
-      nestedCaptures,
-      tdzFlaggedNested,
-    );
-    return { kind: "ref", typeIdx: structTypeIdx };
-  }
-
-  const userParams = sig.params;
-
-  const wrapperTypes = getOrCreateFuncRefWrapperTypes(ctx, userParams, sig.results);
-  if (!wrapperTypes) return null;
-
-  const { structTypeIdx, liftedFuncTypeIdx, closureInfo } = wrapperTypes;
-
-  // Create a trampoline function for THIS specific function.
-  // The trampoline takes (self, ...userParams) and calls the original function.
-  const trampolineName = `__fn_tramp_${funcName}_${ctx.closureCounter++}`;
-  const trampolineBody: Instr[] = [];
-
-  // Push the user-visible params (skip self at param 0)
-  for (let i = 0; i < userParams.length; i++) {
-    trampolineBody.push({ op: "local.get", index: i + 1 });
-  }
-  trampolineBody.push({ op: "call", funcIdx });
-
-  const trampolineFuncIdx = mintDefinedFunc(ctx);
-  pushDefinedFunc(ctx, trampolineFuncIdx, {
-    name: trampolineName,
-    typeIdx: liftedFuncTypeIdx,
-    locals: [],
-    body: trampolineBody,
-    exported: false,
-  });
-  ctx.funcMap.set(trampolineName, trampolineFuncIdx);
-
-  // Emit: ref.func $trampoline, struct.new $closure_struct
-  fctx.body.push({ op: "ref.func", funcIdx: trampolineFuncIdx });
-  fctx.body.push({ op: "struct.new", typeIdx: structTypeIdx });
-
-  return { kind: "ref", typeIdx: structTypeIdx };
 }
 
 /**
